requirements.txt
Django==5.0.1
psycopg2==2.9.9
Pillow==10.2.0
dadata==21.10.1
yookassa==3.0.0
celery[redis]==5.3.6
flower==2.0.1
django-simple-captcha==0.6.0
pyTelegramBotAPI==4.15.3
environs~=10.3.0
aiohttp==3.9.2
urllib3~=2.1.0
requests~=2.31.0
django-debug-toolbar==4.2.0
